backend-project/requirements.txt
fastapi[standard]
pydantic[email]
cryptography
databases
psycopg2-binary
PyMySQL
SQLAlchemy
uvicorn[standard]
gunicorn
alembic
python-multipart
pillow
numpy
openai
stanza
pytest
langchain
language-tool-python
transformers
torch
streamlit
pandas
scipy
matplotlib
webdataset
python-jose[cryptography]
passlib[bcrypt]
pylti
opencv-python
scikit-image
celery[redis]
flower
gevent
memory_profiler
psutil
pytest_asyncio
websockets